+++ example/lib/main.dart
@@ -34,10 +34,12 @@ class _MyHomePageState extends State<MyHomePage> {
           titleColor: Colors.blue,
           title: 'Container Title',
           fontSize: 16.0,
+          backgroundColor: Colors.white,
           child: Container(
             width: 250.0,
             height: 200.0,
             decoration: BoxDecoration(
+              color: Colors.white,
               border: Border.all(
                 color: Colors.blue,
               ),